+++ jopa-impl/src/main/java/cz/cvut/kbss/jopa/model/metamodel/EntityLifecycleListenerManager.java
@@ -1,16 +1,14 @@
 /**
  * Copyright (C) 2016 Czech Technical University in Prague
- *
- * This program is free software: you can redistribute it and/or modify it under
- * the terms of the GNU General Public License as published by the Free Software
- * Foundation, either version 3 of the License, or (at your option) any
- * later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
- * details. You should have received a copy of the GNU General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ * <p>
+ * This program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public
+ * License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later
+ * version.
+ * <p>
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied
+ * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+ * details. You should have received a copy of the GNU General Public License along with this program. If not, see
+ * <http://www.gnu.org/licenses/>.
  */
 package cz.cvut.kbss.jopa.model.metamodel;
 
@@ -72,18 +70,17 @@ public class EntityLifecycleListenerManager {
             parent.invokeEntityListenerCallbacks(instance, lifecycleEvent);
         }
         if (entityListeners != null) {
-            entityListeners.forEach(listener -> {
-                getEntityListenerCallback(listener, lifecycleEvent).ifPresent(method -> {
-                    if (!method.isAccessible()) {
-                        method.setAccessible(true);
-                    }
-                    try {
-                        method.invoke(listener, instance);
-                    } catch (IllegalAccessException | InvocationTargetException e) {
-                        throw new OWLPersistenceException("Unable to invoke entity listener method " + method, e);
-                    }
-                });
-            });
+            entityListeners
+                    .forEach(listener -> getEntityListenerCallback(listener, lifecycleEvent).ifPresent(method -> {
+                        if (!method.isAccessible()) {
+                            method.setAccessible(true);
+                        }
+                        try {
+                            method.invoke(listener, instance);
+                        } catch (IllegalAccessException | InvocationTargetException e) {
+                            throw new OWLPersistenceException("Unable to invoke entity listener method " + method, e);
+                        }
+                    }));
         }
     }