+++ ontodriver-jena/src/test/java/cz/cvut/kbss/ontodriver/jena/connector/InferenceConnectorFactoryTest.java
@@ -17,7 +17,7 @@ public class InferenceConnectorFactoryTest extends ConnectorFactoryTestBase {
 
     @Override
     SharedStorageConnector getCentralConnector(ConnectorFactory factory) throws Exception {
-        final Field connectorField = InferenceConnectorFactory.class.getDeclaredField("centralConnector");
+        final Field connectorField = SharedConnectorBasedConnectorFactory.class.getDeclaredField("centralConnector");
         connectorField.setAccessible(true);
         return (SharedStorageConnector) connectorField.get(factory);
     }